+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### v5.x
 
-- Add more crypto utils
+- Add more crypto/codec utils
 - Remove features related to `no_std` and `wasm32`
 
 #### v4.x